+++ ui/dashboard/src/views/Login.vue
@@ -3,95 +3,15 @@
     <div
       class="column is-three-quarters-mobile is-two-thirds-tablet is-half-desktop is-one-third-widescreen is-one-quarter-fullhd"
     >
-      <div class="card">
-        <header class="card-header">
-          <p class="card-header-title">Login</p>
-        </header>
-        <div class="card-content">
-          <div class="content">
-            <div class="field">
-              <p class="control has-icons-left">
-                <input
-                  class="input"
-                  v-model="username"
-                  placeholder="Username"
-                  required
-                />
-                <span class="icon is-small is-left">
-                  <i class="fas fa-user"></i>
-                </span>
-              </p>
-            </div>
-            <div class="field">
-              <p class="control has-icons-left has-icons-right">
-                <input
-                  class="input"
-                  :type="showPassword ? 'text' : 'password'"
-                  placeholder="Password"
-                  v-model="password"
-                  required
-                />
-                <span class="icon is-small is-left">
-                  <i class="fas fa-lock"></i>
-                </span>
-                <a
-                  class="icon is-small is-right"
-                  style="pointer-events: all"
-                  @click="toggleShowPassword"
-                >
-                  <i class="fas fa-eye"></i>
-                </a>
-              </p>
-            </div>
-          </div>
-        </div>
-        <footer class="card-footer">
-          <a
-            class="card-footer-item button is-white has-text-primary"
-            :class="{ 'is-loading': loading }"
-            @click="submit"
-          >
-            Login&nbsp;
-            <span class="icon is-medium">
-              <i class="fas fa-chevron-right"></i>
-            </span>
-          </a>
-        </footer>
-      </div>
+      <LoginForm />
     </div>
   </div>
 </template>
 
 <script>
+import LoginForm from "@/components/LoginForm";
 export default {
-  data() {
-    return {
-      showPassword: false,
-      username: "",
-      password: "",
-      loading: false,
-    };
-  },
-  methods: {
-    toggleShowPassword() {
-      this.showPassword = !this.showPassword;
-    },
-    submit() {
-      const user = {
-        username: this.username,
-        password: this.password,
-      };
-      this.loading = true;
-      this.$store
-        .dispatch("auth/login", user)
-        .then(() => {
-          this.$router.push({ name: "Dashboard" });
-        })
-        .finally(() => {
-          this.loading = false;
-        });
-    },
-  },
+  components: { LoginForm },
 };
 </script>
